Fixed RemoveFromView and moongate teleport regressions. Fixed a couple more undefined behaviors.

+++ src/game/clients/CClientMsg_AOSTooltip.cpp
@@ -16,7 +16,7 @@ uint HashString(lpctstr str, size_t length)
 {
 	uint hash = 5381;
 	for (size_t i = 0; i < length; ++i)
-		hash = ((hash << 5) + hash) + *(str++);
+		hash = ((hash << 5) + hash) + (uint)(*(str++));
 
 	return hash;
 }
@@ -111,7 +111,7 @@ bool CClient::addAOSTooltip(CObjBase * pObj, bool fRequested, bool fShop)
 
                 pObj->AddPropsTooltipData(pObj);
 			}
-			
+
 			if (IsTrigUsed(TRIGGER_CLIENTTOOLTIP_AFTERDEFAULT) || (pItem && IsTrigUsed(TRIGGER_ITEMCLIENTTOOLTIP_AFTERDEFAULT)) || (pChar && IsTrigUsed(TRIGGER_CHARCLIENTTOOLTIP_AFTERDEFAULT)))
 			{
 				CScriptTriggerArgs args(pObj);
@@ -165,7 +165,7 @@ bool CClient::addAOSTooltip(CObjBase * pObj, bool fRequested, bool fShop)
 			pObj->SetPropertyList(propertyList);
 		}
 	}
-	
+
 	if (propertyList->isEmpty() == false)
 	{
 		switch (g_Cfg.m_iTooltipMode)
@@ -661,7 +661,7 @@ void CClient::AOSTooltip_addDefaultItemData(CItem * pItem)
 		t->FormatArgs("Time range\t%hu min / %hu max", pSpawn->GetTimeLo(), pSpawn->GetTimeHi());
 		PUSH_BACK_TOOLTIP(pItem, t = new CClientTooltip(1060660)); // ~1_val~: ~2_val~
 		t->FormatArgs("Time until next spawn\t%" PRId64 " sec", pItem->GetTimerSAdjusted());
-		
+
 	} break;
 
 	case IT_SPAWN_ITEM:
